% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_addin.R
\name{syn_gadget}
\alias{syn_gadget}
\title{Synonym Antonym Gadget}
\usage{
syn_gadget(word, synonym = TRUE)
}
\arguments{
\item{word}{A single word to look up}

\item{synonym}{Show synonym (`TRUE`) or antonym (`FALSE`)}
}
\description{
Gadget that powers synonym/antonym replacement.
}
