% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_addin.R
\name{syn_addin}
\alias{syn_addin}
\title{Synonym and Antonym Replacement Addin}
\usage{
syn_addin(synonym = TRUE)
}
\description{
Gets currently selected word, searches for synonym or antonym using [syn],
allows user to pick replacement.
}
